pyproject.toml
[project]
name = "Vortek"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Magno", email = "dutracm04@gmail.com"},
    {name = "Anteconix", email = "dutracm04@gmail.com"},
]
dependencies = [
    "pdm-vscode>=0.1.1",
    "pdm-autoexport>=0.2.0",
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.2.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
